feature/github-actions -- temp renamed Input to InputComponent
