+++ essential-ui/src/components/data-entry/InputComponent/main/view.tsx
@@ -1,4 +1,4 @@
-// @TODO -- Implement the Input component
+// @TODO -- Implement the Input component.
 export function Input() {
   return <input></input>;
 }
